back-end/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pandas
scikit-learn==1.6.1
joblib
matplotlib
seaborn
requests
pydantic[email]
Pillow
python-dotenv
python-multipart
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
bcrypt
psycopg2-binary
email-validator
